Add files via upload
This dashboard provides insights into the data professional landscape, based on a survey of data professionals. It shows the average salary by job title, the most popular programming languages, the difficulty of breaking into the data field, and the country of residence of survey takers. This information can be helpful for data professionals who are looking to learn more about the industry or make career decisions.