Add files via upload
This dashboard visualizes data on the Airbnb market in Seattle. It shows the average price per bedroom, the distinct count of listings per bedroom, and the price per zip code. The dashboard also shows the total revenue for the year. This information can be helpful for people who are interested in the Airbnb market in Seattle.